fix(agents): retain prepared video providers through background jobs
