+++ src/agents/tools/media-generation-tool-providers.ts
@@ -6,7 +6,10 @@ import { withPluginRuntimeGenerationScope } from "../../plugins/runtime/generati
 import { AsyncWorkScope } from "../../shared/async-work-scope.js";
 import type { PreparedModelRuntimeSnapshot } from "../prepared-model-runtime.types.js";
 
-type MediaProviderKey = "imageGenerationProviders" | "musicGenerationProviders";
+type MediaProviderKey =
+  | "imageGenerationProviders"
+  | "musicGenerationProviders"
+  | "videoGenerationProviders";
 type MediaProviderOptions = { cfg: OpenClawConfig; prepared?: PreparedModelRuntimeSnapshot };
 
 export function acquireImageGenerationToolProviders(params: MediaProviderOptions) {
@@ -17,11 +20,19 @@ export function acquireMusicGenerationToolProviders(params: MediaProviderOptions
   return acquireMediaGenerationToolProviders("musicGenerationProviders", params);
 }
 
+export function acquireVideoGenerationToolProviders(params: MediaProviderOptions) {
+  return acquireMediaGenerationToolProviders("videoGenerationProviders", params);
+}
+
 async function acquireMediaGenerationToolProviders<K extends MediaProviderKey>(
   key: K,
   params: MediaProviderOptions,
 ) {
-  const label = key === "imageGenerationProviders" ? "Image" : "Music";
+  const label = {
+    imageGenerationProviders: "Image",
+    musicGenerationProviders: "Music",
+    videoGenerationProviders: "Video",
+  }[key];
   const work = new AsyncWorkScope();
   const prepared = params.prepared;
   const inGeneration = <T>(run: () => T): T =>
